server/requirements.txt
amqp==2.5.2
Authlib==0.14.2
Babel==2.8.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
entrypoints==0.3
Faker==4.0.3
flake8==3.7.9
Flask==1.1.2
Flask-Avatars==0.2.2
Flask-Babel==1.0.0
Flask-CKEditor==0.4.4.1
Flask-Cors==3.0.8
Flask-DebugToolbar==0.11.0
Flask-Dropzone==1.5.4
Flask-HTTPAuth==4.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.12.0
Flask-Moment==0.9.0
flask-redis==0.4.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
flask-whooshee==0.7.0
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
itchat==1.3.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
lock==2018.3.25.2110
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyMySQL==0.9.3
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-engineio==3.12.1
python-socketio==4.5.1
python-sql==1.1.0
pytz==2020.1
redis==3.5.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
text-unidecode==1.3
urllib3==1.25.9
vine==1.3.0
webargs==6.1.0
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.1
zipp==3.1.0
